Updated .gitignore file to ignore all node_modules, created and started all handlebar files
