+++ Develop/controllers/homeRoutes.js
@@ -1,0 +1,4 @@
+
+const router = require('express').Router();
+
+module.exports = router;
