+++ Develop/views/partials/book-search-page.handlebars
@@ -1,0 +1,37 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+  <meta charset="UTF-8">
+
+  <meta name="viewport" 
+  content="width=device-width, initial-scale=1.0">
+    <title>If Books Could Kill</title>
+    
+     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css" 
+  rel="stylesheet" 
+  integrity="sha384-T3c6CoIi6uLrA9TneNEoa7RxnatzjcDSCmG1MXxSR1GAsXEV/Dwwykc2MPK8M2HN" 
+  crossorigin="anonymous">
+
+</head>
+<body>
+
+<nav class="navbar">
+    <div class="container">
+        <div class="navbar-header">
+            <a class="navbar-brand" href="#">If Books Could Kill</a>
+        </div>
+
+        <div class="navbar-search">
+            <form class="form-inline">
+                <input class="form-control" type="text" placeholder="Search Books">
+                <button class="btn btn-default" type="submit">Search</button>
+            </form>
+        </div>
+    </div>
+</nav>
+
+<!-- The rest of your book search content goes here -->
+
+</body>
+</html>
