+++ Develop/views/partials/checkout-page.handlebars
@@ -1,0 +1,24 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <title>Checkout...If you dare...</title>
+    <style>
+        body {
+            text-align: center;
+            font-family: Arial, sans-serif;
+        }
+        .title {
+            font-size: 24px;
+            margin-top: 20px;
+        }
+    </style>
+</head>
+<body>
+    <div class="title">
+        Checkout...If you dare...
+    </div>
+
+    <!-- Your checkout form and content go here -->
+
+</body>
+</html>
