+++ Develop/views/partials/login-page.handlebars
@@ -1,0 +1,46 @@
+<div class="row">
+  <div class="col-md-6">
+    <h2>Login</h2>
+
+    <form class="form login-form">
+      <div class="form-group">
+        <label for="email-login">Email:</label>
+        <input class="form-input" type="text" id="email-login" />
+      </div>
+      <div class="form-group">
+        <label for="password-login">Password:</label>
+        <input class="form-input" type="password" id="password-login" />
+      </div>
+      <div class="form-group">
+        <button class="btn btn-primary" type="submit">Login</button>
+      </div>
+    </form>
+  </div>
+  <div class="col-md-6">
+    <h2>Create Account</h2>
+
+    <form class="form sign-up-form">
+      <div class="form-group">
+        <label for="first-name-sign-up">First Name:</label>
+        <input class="form-input" type="text" id="first-name-sign-up" />
+      </div>
+      <div class="form-group">
+        <label for="last-name-sign-up">Last Name:</label>
+        <input class="form-input" type="text" id="last-name-sign-up" />
+      </div>
+      <div class="form-group">
+        <label for="email-sign-up">Email:</label>
+        <input class="form-input" type="text" id="email-sign-up" />
+      </div>
+      <div class="form-group">
+        <label for="password-sign-up">Password:</label>
+        <input class="form-input" type="password" id="password-sign-up" />
+      </div>
+      <div class="form-group">
+        <button class="btn btn-primary" type="submit">Create Account</button>
+      </div>
+    </form>
+  </div>
+</div>
+
+<script src="./js/login.js"></script>
