+++ package.json
@@ -1,0 +1,8 @@
+{
+  "dependencies": {
+    "bootstrap": "^5.3.2",
+    "express": "^4.18.2",
+    "express-session": "^1.17.3",
+    "handlebars.moment": "^1.0.4"
+  }
+}
